requirement.txt
spacy==3.6.0
ipython==8.8.0
torch==2.0.1
transformers==4.30.2
tqdm==4.65.0
nltk==3.8.1
PyYAML==6.0
joblib==1.3.1
DateTime==5.1
numpy==1.23.5
pandas==2.0.3